+++ src/commands/openCommitInRemote.ts
@@ -5,10 +5,10 @@ import { ActiveEditorCommand, Commands } from './commands';
 import { GitCommit, GitService, GitUri } from '../gitService';
 import { Logger } from '../logger';
 
-export class OpenCommitInHostingProviderCommand extends ActiveEditorCommand {
+export class OpenCommitInRemoteCommand extends ActiveEditorCommand {
 
     constructor(private git: GitService, private repoPath: string) {
-        super(Commands.OpenCommitInHostingProvider);
+        super(Commands.OpenCommitInRemote);
     }
 
     async execute(editor: TextEditor, edit: TextEditorEdit, uri?: Uri) {
@@ -27,7 +27,7 @@ export class OpenCommitInHostingProviderCommand extends ActiveEditorCommand {
             if (blameline < 0) return undefined;
 
             const blame = await this.git.getBlameForLine(gitUri, blameline);
-            if (!blame) return window.showWarningMessage(`Unable to open commit in hosting provider. File is probably not under source control`);
+            if (!blame) return window.showWarningMessage(`Unable to open commit in remote provider. File is probably not under source control`);
 
             let commit = blame.commit;
             // If the line is uncommitted, find the previous commit
@@ -36,11 +36,11 @@ export class OpenCommitInHostingProviderCommand extends ActiveEditorCommand {
             }
 
             const remotes = Arrays.uniqueBy(await this.git.getRemotes(this.repoPath), _ => _.url, _ => !!_.provider);
-            return commands.executeCommand(Commands.OpenInHostingProvider, uri, remotes, 'commit', [commit.sha]);
+            return commands.executeCommand(Commands.OpenInRemote, uri, remotes, 'commit', [commit.sha]);
         }
         catch (ex) {
-            Logger.error('[GitLens.OpenCommitInHostingProviderCommand]', ex);
-            return window.showErrorMessage(`Unable to open commit in hosting provider. See output channel for more details`);
+            Logger.error('[GitLens.OpenCommitInRemoteCommand]', ex);
+            return window.showErrorMessage(`Unable to open commit in remote provider. See output channel for more details`);
         }
     }
 }