+++ src/git/remotes/factory.ts
@@ -1,28 +1,28 @@
 'use strict';
-import { HostingProvider } from './hostingProvider';
+import { RemoteProvider } from './provider';
 import { GitHubService } from './github';
 import { Logger } from '../../logger';
 
-export { HostingProvider };
+export { RemoteProvider };
 
-const serviceMap = new Map<string, (domain: string, path: string) => HostingProvider>([
+const providerMap = new Map<string, (domain: string, path: string) => RemoteProvider>([
     ['github.com', (domain: string, path: string) => new GitHubService(domain, path)]
 ]);
 
 const UrlRegex = /^(?:git:\/\/(.*?)\/|https:\/\/(.*?)\/|http:\/\/(.*?)\/|git@(.*):\/\/|ssh:\/\/git@(.*?)\/)(.*)$/;
 
-export class HostingProviderFactory {
+export class RemoteProviderFactory {
 
-    static getHostingProvider(url: string): HostingProvider {
+    static getRemoteProvider(url: string): RemoteProvider {
         try {
             const match = UrlRegex.exec(url);
             const domain = match[1] || match[2] || match[3] || match[4] || match[5];
             const path = match[6].replace(/\.git/, '');
 
-            const serviceCreator = serviceMap.get(domain);
-            if (!serviceCreator) return undefined;
+            const creator = providerMap.get(domain);
+            if (!creator) return undefined;
 
-            return serviceCreator(domain, path);
+            return creator(domain, path);
         }
         catch (ex) {
             Logger.error(ex);